Time: 110 ms (64.64%) | Memory: 11.3 MB (49.73%) - LeetSync

+++ 1890-sum-of-beauty-of-all-substrings/sum-of-beauty-of-all-substrings.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int beautySum(string s) {
         int n = s.size();
-        long long ans = 0;
+        int ans = 0;
         for(int i=0;i<n;i++){
             vector<int>freq(26,0);
